trigger the RemoveFavorite confirmation dialog on the favorites dashboard UI
