+++ app/src/main/kotlin/org/cru/godtools/ui/tools/ToolsListFragment.kt
@@ -1,15 +1,12 @@
 package org.cru.godtools.ui.tools
 
-import android.app.Dialog
 import android.graphics.drawable.NinePatchDrawable
 import android.os.Bundle
 import androidx.annotation.CallSuper
 import androidx.core.content.ContextCompat
-import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.LifecycleOwner
 import androidx.recyclerview.widget.RecyclerView
-import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.h6ah4i.android.widget.advrecyclerview.animator.DraggableItemAnimator
 import com.h6ah4i.android.widget.advrecyclerview.draggable.RecyclerViewDragDropManager
 import com.h6ah4i.android.widget.advrecyclerview.utils.WrapperAdapterUtils
@@ -32,9 +29,7 @@ import org.cru.godtools.analytics.model.AnalyticsScreenEvent.Companion.SCREEN_AL
 import org.cru.godtools.analytics.model.AnalyticsScreenEvent.Companion.SCREEN_HOME
 import org.cru.godtools.analytics.model.AnalyticsScreenEvent.Companion.SCREEN_LESSONS
 import org.cru.godtools.base.Settings
-import org.cru.godtools.base.ui.util.getName
 import org.cru.godtools.databinding.ToolsFragmentBinding
-import org.cru.godtools.download.manager.GodToolsDownloadManager
 import org.cru.godtools.fragment.BasePlatformFragment
 import org.cru.godtools.model.Tool
 import org.cru.godtools.model.Translation
@@ -45,7 +40,6 @@ import org.cru.godtools.tutorial.analytics.model.TUTORIAL_HOME_DISMISS
 import org.cru.godtools.tutorial.analytics.model.TutorialAnalyticsActionEvent
 import org.cru.godtools.widget.BannerType
 import org.keynote.godtools.android.db.GodToolsDao
-import splitties.fragmentargs.arg
 import splitties.fragmentargs.argOrDefault
 
 @AndroidEntryPoint
@@ -68,8 +62,6 @@ class ToolsListFragment() : BasePlatformFragment<ToolsFragmentBinding>(R.layout.
 
     @Inject
     internal lateinit var dao: GodToolsDao
-    @Inject
-    internal lateinit var downloadManager: GodToolsDownloadManager
 
     // region Lifecycle
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -184,17 +176,6 @@ class ToolsListFragment() : BasePlatformFragment<ToolsFragmentBinding>(R.layout.
     }
     // endregion ToolsAdapterCallbacks
 
-    // region Remove Favorite
-    private fun showRemoveFavoriteConfirmationDialog(tool: Tool?, translation: Translation?) {
-        RemoveFavoriteConfirmationDialogFragment(tool?.code ?: return, translation.getName(tool, null).toString())
-            .show(childFragmentManager, null)
-    }
-
-    internal fun removeFavorite(code: String?) {
-        code?.let { downloadManager.unpinToolAsync(it) }
-    }
-    // endregion Remove Favorite
-
     // region Tools List
     private val toolsAdapterDataModel by viewModels<ToolsAdapterViewModel>()
     private val toolsAdapter: ToolsAdapter by lazy {
@@ -262,24 +243,4 @@ class ToolsListFragment() : BasePlatformFragment<ToolsFragmentBinding>(R.layout.
         toolsDragDropAdapter = null
     }
     // endregion Tools List
-}
-
-class RemoveFavoriteConfirmationDialogFragment() : DialogFragment() {
-    constructor(code: String, name: String) : this() {
-        this.code = code
-        this.name = name
-    }
-
-    private var code: String by arg()
-    private var name: String by arg()
-
-    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
-        return MaterialAlertDialogBuilder(requireContext())
-            .setTitle(getString(R.string.tools_list_remove_favorite_dialog_title, name))
-            .setPositiveButton(R.string.tools_list_remove_favorite_dialog_confirm) { _, _ ->
-                findListener<ToolsListFragment>()?.removeFavorite(code)
-            }
-            .setNegativeButton(R.string.tools_list_remove_favorite_dialog_dismiss, null)
-            .create()
-    }
 }
